check for previous instance of user file

+++ commands/join.js
@@ -6,22 +6,26 @@ module.exports = {
 		.setName('join')
 		.setDescription('Join the family, get your first car, and get ready to race!'),
 	async execute(interaction) {
-        newUser = {
-            "id": interaction.user.id,
-            "name": interaction.user.name,
-            "bal": 0,
-            "current_car_id": "000",
-            "cars": [],
-            "items": [],
-            "mods": []
+        if (fs.existsSync(`./data/users/${interaction.user.id}.json`)){
+            await interaction.reply(`You're already part of the family <@${interaction.user.id}>!`);
+        } else {
+            newUser = {
+                "id": interaction.user.id,
+                "name": interaction.user.name,
+                "bal": 0,
+                "current_car_id": "000",
+                "cars": [],
+                "items": [],
+                "mods": []
+            }
+            
+            var jsonString = JSON.stringify(newUser);
+    
+            // Probably want to async this at some point
+            fs.writeFileSync(`./data/users/${interaction.user.id}.json`, jsonString);
+            console.log(`${interaction.user.id}.json created`);
+    
+            await interaction.reply(`Welcome to the family <@${interaction.user.id}>! You\'ll need a car before you can get yourself some Coronas, which one do you wanna start with?`);
         }
-        
-        var jsonString = JSON.stringify(newUser);
-
-        // Probably want to async this at some point
-        fs.writeFileSync(`./data/users/${interaction.user.id}.json`, jsonString);
-        console.log(`${interaction.user.id}.json created`);
-
-		await interaction.reply(`Welcome to the family <@${interaction.user.id}>! You\'ll need a car before you can get yourself some Coronas, which one do you wanna start with?`);
 	},
 };
